pyproject.toml
[tool.poetry]
name = "reprosyn"
version = "0.1.0"
description = ""
authors = ["callummole <cmole@turing.ac.uk>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
private-pgm = {git = "https://github.com/ryan112358/private-pgm.git"}
disjoint-set = "^0.7.3"
numpy = "^1.22.3"
pandas = "^1.4.2"
jaxlib = { version = "^0.3.15", optional = true }
jax = { version = "^0.3.16", optional = true }
tqdm = "^4.64.0"
DataSynthesizer = "^0.1.11"
tensorflow = "^2.10.0"
ektelo = { git = "https://github.com/callummole/ektelo.git", rev = "dependencies", optional = true }
py-synthpop = "^0.1.2"
validators = "^0.20.0"
ctgan = {git = "https://github.com/alan-turing-institute/CTGAN.git", rev = "dependencies"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
jupyter = "^1.0.0"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pyinstaller = "^5.0.1"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.3.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-click = "^4.3.0"
myst-parser = "^0.18.1"


[tool.poetry.extras]
jax = ["jax","jaxlib"] #see issues #11 and #50
ektelo = ["ektelo"] #see issue #53

[tool.poetry.scripts]
rsyn = "reprosyn.cli:cli"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
